+++ saleor/app/management/commands/install_app.py
@@ -53,10 +53,9 @@ class Command(BaseCommand):
             app_job.permissions.set(permissions)
 
         try:
-            app = install_app(app_job, activate)
+            _app, token = install_app(app_job, activate)
         except Exception as e:
             app_job.status = JobStatus.FAILED
             app_job.save()
             raise e
-        token = app.tokens.first()
-        return json.dumps({"auth_token": token.auth_token})
+        return json.dumps({"auth_token": token})