+++ saleor/graphql/account/types.py
@@ -9,6 +9,7 @@ from ...checkout.utils import get_user_checkout
 from ...core.exceptions import PermissionDenied
 from ...core.permissions import AccountPermissions, OrderPermissions
 from ...core.tracing import traced_resolver
+from ...order import OrderStatus
 from ...order import models as order_models
 from ..checkout.dataloaders import CheckoutByUserAndChannelLoader, CheckoutByUserLoader
 from ..checkout.types import Checkout
@@ -19,9 +20,12 @@ from ..core.scalars import UUID
 from ..core.types import CountryDisplay, Image, Permission
 from ..core.utils import from_global_id_or_error, str_to_enum
 from ..decorators import one_of_permissions_required, permission_required
+from ..giftcard.dataloaders import GiftCardsByUserLoader
 from ..meta.types import ObjectWithMetadata
-from ..utils import format_permissions_for_display
+from ..order.dataloaders import OrdersByUserLoader
+from ..utils import format_permissions_for_display, get_user_or_app_from_context
 from ..wishlist.resolvers import resolve_wishlist_items_from_user
+from .dataloaders import CustomerEventsByUserLoader
 from .enums import CountryCodeEnum, CustomerEventsEnum
 from .utils import can_user_manage_group, get_groups_which_user_can_manage
 
@@ -303,7 +307,7 @@ class User(CountableDjangoObjectType):
     @staticmethod
     @traced_resolver
     def resolve_gift_cards(root: models.User, info, **_kwargs):
-        return root.gift_cards.all()
+        return GiftCardsByUserLoader(info.context).load(root.id)
 
     @staticmethod
     @traced_resolver
@@ -335,15 +339,18 @@ class User(CountableDjangoObjectType):
     )
     @traced_resolver
     def resolve_events(root: models.User, info):
-        return root.events.all()
+        return CustomerEventsByUserLoader(info.context).load(root.id)
 
     @staticmethod
     @traced_resolver
     def resolve_orders(root: models.User, info, **_kwargs):
-        viewer = info.context.user
-        if viewer.has_perm(OrderPermissions.MANAGE_ORDERS):
-            return root.orders.all()
-        return root.orders.non_draft()  # type: ignore
+        def _resolve_orders(orders):
+            requester = get_user_or_app_from_context(info.context)
+            if requester.has_perm(OrderPermissions.MANAGE_ORDERS):
+                return orders
+            return list(filter(lambda order: order.status != OrderStatus.DRAFT, orders))
+
+        return OrdersByUserLoader(info.context).load(root.id).then(_resolve_orders)
 
     @staticmethod
     @traced_resolver