+++ saleor/graphql/app/tests/mutations/test_app_token_verify.py
@@ -1,3 +1,4 @@
+from .....app.models import AppToken
 from ....tests.utils import get_graphql_content
 
 APP_TOKEN_VERIFY_MUTATION = """
@@ -10,7 +11,7 @@ mutation AppTokenVerify($token: String!){
 
 
 def test_app_token_verify_valid_token(app, api_client):
-    token = app.tokens.first().auth_token
+    _token_obj, token = AppToken.objects.create_app_token(app=app)
     query = APP_TOKEN_VERIFY_MUTATION
 
     variables = {"token": token}
@@ -19,8 +20,21 @@ def test_app_token_verify_valid_token(app, api_client):
     assert content["data"]["appTokenVerify"]["valid"]
 
 
+def test_app_token_verify_valid_token_and_inactive_app(app, api_client):
+    app.is_active = False
+    app.save(update_fields=["is_active"])
+
+    _token_obj, token = AppToken.objects.create_app_token(app=app)
+    query = APP_TOKEN_VERIFY_MUTATION
+
+    variables = {"token": token}
+    response = api_client.post_graphql(query, variables=variables)
+    content = get_graphql_content(response)
+    assert not content["data"]["appTokenVerify"]["valid"]
+
+
 def test_app_token_verify_invalid_token(app, api_client):
-    token = app.tokens.first().auth_token
+    _token_obj, token = AppToken.objects.create_app_token(app=app)
     token += "incorrect"
     query = APP_TOKEN_VERIFY_MUTATION